Add methods to ClockTree to improve performances
Computing the clock tree requires a lot of read access to registers.
By default, the cache is disabled so every register access are made to
device which is not efficient. Worse, the registers could change while
we are processing the clock tree.
Two avoid that, some have been added to ClockTree to read all registers
at once, and use the cache.

Note that ClockTree doesn't use them.
The methods are expected to be used by application at place where it make
sense to use them.

Signed-off-by: Alexandre Bailon <alexandre.bailon@gmail.com>